requirements.txt
fastapi
uvicorn
python-multipart
torch
transformers
pillow
opencv-python-headless
numpy
